Update activity panels
